Refactor geotechnical data processing and visualization utilities
- Removed obsolete scripts for separating CPT sheets, combining CPT and SPT data, and merging SPT-geo combined data with shearbox data.
- Introduced new classes: GeoDataCombiner and GeoDataSeparator for better organization of data processing functions.
- Added GeoProfileVisualizer class for visualizing borehole and SPT profiles.
- Updated methods for loading data, preparing depth and code mappings, and plotting profiles.
- Enhanced modularity and readability of the codebase by consolidating related functionalities into dedicated classes.
