+++ src/__init__.py
@@ -7,8 +7,20 @@ from .modeling import (
     NNVoxelModel,
     StackedVoxelModel,
 )
+
 from .borehole import (
     AGSProcessor,
     cleanse_borehole_data,
     process_ags_geology,
+)
+
+from .utils import (
+    GeoDataCombiner,
+    GeoDataSeparator,
+    GeoProfileVisualizer,
+)
+
+from .comparison import (
+    compare_borehole_logs,
+    SectionComparison,
 )
