+++ src/comparison/__init__.py
@@ -1,0 +1,2 @@
+from .borehole_log_compare import compare_borehole_logs
+from .cross_section_compare import SectionComparison
