+++ src/utils/visualization/__init__.py
@@ -1,0 +1,1 @@
+from .profile_visual import GeoProfileVisualizer
